Use newest version of WK Item Info Injector
The newest version adds support for the new accordion UI during the lesson quiz (when script compatibility mode is disabled). For this, WK Item Info Injector uses the heading argument provided to `append()` as the title of the collapsed section. Since Anime Sentences passes `null`, this results in an empty heading. So here is a pull request with a suggestion how to make Anime Sentences more compatible to the new version of WK Item Info Injector.